requirements.txt
diffusers
dreamsim
fairscale
opencv_python
opencv_python_headless
pillow
peft
requests
sentencepiece
sentence_transformers
setuptools
scikit-image
submitit
tensorboardX
timm
torch>=2.4.0
torchvision>=0.19.0
tqdm
transformers
vllm